Fix endless loop in RedisJsonModel and other redis fixes
Use redis `exists` command correctly so return value is handled correctly. I'm certain this was a boolean return value previously but must have changed to an integer and some point. Was making all exists calls look true and causing the endless loop on the tin.

Fix a warning about redis `multi` usage needing to use the instance yielded to the block. Lock redis version to <5 to fix errors from actioncable.